+++ changelog.md
@@ -1,5 +1,4 @@
 ## General Changes
-- Code cleanup and optimizations
-- Fixed folia support issues
+- Added a playerutils class to handle player-related operations more efficiently & supported for folia.
 
-**Full Changelog**: https://github.com/yL3oft/zAPI/compare/1.4.3...1.4.4
+**Full Changelog**: https://github.com/yL3oft/zAPI/compare/1.4.4...1.4.5